Refactoring
Moved CRON web interface pages to sub-folder.
