+++ webfiori/framework/cron/webUI/CronTasksView.php
@@ -22,7 +22,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-namespace webfiori\framework\cron;
+namespace webfiori\framework\cron\webUI;
 
 use webfiori\framework\File;
 use webfiori\framework\WebFioriApp;
